requirements.txt
fastapi
uvicorn[standard]
pandas
plotly
requests
prophet
pytorch-forecasting[lightning]
torch
python-dotenv
beautifulsoup4
darts